+++ id/id/src/main/java/io/piano/android/id/models/PianoIdApi.kt
@@ -6,6 +6,7 @@ import retrofit2.http.Field
 import retrofit2.http.FieldMap
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.GET
+import retrofit2.http.Header
 import retrofit2.http.POST
 import retrofit2.http.Query
 import retrofit2.http.Url
@@ -15,7 +16,7 @@ internal interface PianoIdApi {
     fun signOut(
         @Url url: String,
         @Query("client_id") aid: String,
-        @Query("token") accessToken: String
+        @Header("Authorization") accessToken: String
     ): Call<ResponseBody>
 
     @POST
@@ -31,6 +32,14 @@ internal interface PianoIdApi {
         @Url url: String,
         @FieldMap tokenParams: Map<String, String>
     ): Call<PianoIdToken>
+
+    @GET
+    fun getUserInfo(
+        @Url url: String,
+        @Query("client_id") aid: String,
+        @Header("Authorization") accessToken: String,
+        @Query("form_name") formName: String?
+    ): Call<PianoUserProfile>
 
     @FormUrlEncoded
     @POST("api/v3/anon/mobile/sdk/id/deployment/host")