+++ id/id/src/test/java/io/piano/android/id/PianoIdClientTest.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.atLeastOnce
 import com.nhaarman.mockitokotlin2.doNothing
@@ -17,6 +18,7 @@ import com.squareup.moshi.Moshi
 import io.piano.android.id.models.HostResponse
 import io.piano.android.id.models.PianoIdApi
 import io.piano.android.id.models.PianoIdToken
+import io.piano.android.id.models.PianoUserProfile
 import okhttp3.HttpUrl
 import okhttp3.ResponseBody
 import retrofit2.Call
@@ -32,9 +34,11 @@ import kotlin.test.assertTrue
 class PianoIdClientTest {
     private val hostResponseCall: Call<HostResponse> = mock()
     private val signOutResponseCall: Call<ResponseBody> = mock()
+    private val userProfileResponseCall: Call<PianoUserProfile> = mock()
     private val api: PianoIdApi = mock {
         on { getDeploymentHost(any()) } doReturn hostResponseCall
         on { signOut(any(), any(), any()) } doReturn signOutResponseCall
+        on { getUserInfo(any(), any(), any(), anyOrNull()) } doReturn userProfileResponseCall
     }
     private val moshi = Moshi.Builder()
         .build()
@@ -249,6 +253,41 @@ class PianoIdClientTest {
             val exc = PianoIdException()
             authUrlCallback(Result.failure(exc))
             val valueCaptor = argumentCaptor<Result<String>>()
+            verify(pianoIdCallback).invoke(valueCaptor.capture())
+            assertTrue { valueCaptor.lastValue.isFailure }
+        }
+
+    private fun getUserInfo(
+        callbackTest: (PianoIdFuncCallback<PianoUserProfile>, PianoIdFuncCallback<HttpUrl>) -> Unit
+    ) {
+        doNothing().`when`(pianoIdClient).getHostUrl(any())
+        val callback: PianoIdFuncCallback<PianoUserProfile> = mock()
+        pianoIdClient.getUserInfo(DUMMY, null, callback)
+        val callbackCaptor = argumentCaptor<PianoIdFuncCallback<HttpUrl>>()
+        verify(pianoIdClient).getHostUrl(callbackCaptor.capture())
+        callbackTest(callback, callbackCaptor.lastValue)
+    }
+
+    @Test
+    fun getUserInfoSuccess() =
+        getUserInfo { pianoIdCallback, authUrlCallback ->
+            authUrlCallback(Result.success(url))
+            verify(api).getUserInfo(any(), any(), any(), anyOrNull())
+            val callbackCaptor = argumentCaptor<Callback<PianoUserProfile>>()
+            verify(userProfileResponseCall).enqueue(callbackCaptor.capture())
+            val response = Response.success<PianoUserProfile>(mock())
+            callbackCaptor.lastValue.onResponse(userProfileResponseCall, response)
+            val valueCaptor = argumentCaptor<Result<PianoUserProfile>>()
+            verify(pianoIdCallback).invoke(valueCaptor.capture())
+            assertTrue { valueCaptor.lastValue.isSuccess }
+        }
+
+    @Test
+    fun getUserInfoFailure() =
+        getUserInfo { pianoIdCallback, authUrlCallback ->
+            val exc = PianoIdException()
+            authUrlCallback(Result.failure(exc))
+            val valueCaptor = argumentCaptor<Result<PianoUserProfile>>()
             verify(pianoIdCallback).invoke(valueCaptor.capture())
             assertTrue { valueCaptor.lastValue.isFailure }
         }